fix settings layout
