+++ apps/desktop/src/components/settings/agents/index.ts
@@ -1,8 +1,6 @@
 export { AgentsSection } from "./AgentsSection";
 export { EmptyLibraryState } from "./EmptyLibraryState";
-export { EmptyTemplatesState } from "./EmptyTemplatesState";
 export { LibraryTab } from "./LibraryTab";
 export { ModelSelect } from "./ModelSelect";
 export { PersonalitySelect } from "./PersonalitySelect";
 export { RoleSelect } from "./RoleSelect";
-export { TemplateCard } from "./TemplateCard";
