+++ apps/desktop/src/components/settings/personalities/PersonalitiesSection.tsx
@@ -18,7 +18,7 @@ import type {
   PersonalityViewModel,
 } from "@fishbowl-ai/ui-shared";
 import { usePersonalitiesStore } from "@fishbowl-ai/ui-shared";
-import { AlertCircle, Plus, RotateCcw, UserPlus, X } from "lucide-react";
+import { AlertCircle, Plus, RotateCcw, X } from "lucide-react";
 import React, { useCallback, useMemo, useState } from "react";
 import { useServices } from "../../../contexts";
 import { cn } from "../../../lib/utils";
@@ -271,9 +271,6 @@ export const PersonalitiesSection: React.FC<PersonalitiesSectionProps> = ({
     if (!isLoading && personalities.length === 0) {
       return (
         <div className="flex flex-col items-center justify-center py-16 px-4">
-          <div className="w-16 h-16 bg-muted rounded-full mx-auto mb-6 flex items-center justify-center">
-            <UserPlus className="h-8 w-8 text-muted-foreground" />
-          </div>
           <h3 className="text-xl font-semibold mb-2 text-center">
             No personalities configured
           </h3>
