+++ apps/desktop/src/components/settings/roles/RolesSection.tsx
@@ -13,19 +13,19 @@
  */
 
 import type {
-  RoleViewModel,
   RoleFormData,
   RolesSectionProps,
+  RoleViewModel,
 } from "@fishbowl-ai/ui-shared";
 import { useRolesStore } from "@fishbowl-ai/ui-shared";
-import React, { useCallback, useState, useMemo } from "react";
-import { UserPlus, Plus, AlertCircle, RotateCcw, X } from "lucide-react";
+import { AlertCircle, Plus, RotateCcw, X } from "lucide-react";
+import React, { useCallback, useMemo, useState } from "react";
+import { useServices } from "../../../contexts";
 import { cn } from "../../../lib/utils";
 import { Button } from "../../ui/button";
-import { RolesList } from "./RolesList";
 import { RoleDeleteDialog } from "./RoleDeleteDialog";
 import { RoleFormModal } from "./RoleFormModal";
-import { useServices } from "../../../contexts";
+import { RolesList } from "./RolesList";
 
 export const RolesSection: React.FC<RolesSectionProps> = ({ className }) => {
   const { logger } = useServices();
@@ -237,9 +237,6 @@ export const RolesSection: React.FC<RolesSectionProps> = ({ className }) => {
     if (!isLoading && roles.length === 0) {
       return (
         <div className="flex flex-col items-center justify-center py-16 px-4">
-          <div className="w-16 h-16 bg-muted rounded-full mx-auto mb-6 flex items-center justify-center">
-            <UserPlus className="h-8 w-8 text-muted-foreground" />
-          </div>
           <h3 className="text-xl font-semibold mb-2 text-center">
             No roles configured
           </h3>
